+++ react/admin/OrganizationDetails.tsx
@@ -37,6 +37,7 @@ import type { Seller } from './OrganizationDetails/OrganizationDetailsSellers'
 import OrganizationDetailsSellers from './OrganizationDetails/OrganizationDetailsSellers'
 import type { PermissionsOptions } from './OrganizationDetails/OrganizationDetailsSettings'
 import OrganizationDetailsSettings from './OrganizationDetails/OrganizationDetailsSettings'
+import { useOrgPermission } from '../hooks/useOrgPermission'
 
 export interface CellRendererProps<RowType> {
   cellData: unknown
@@ -132,6 +133,10 @@ const OrganizationDetails: FunctionComponent = () => {
     variables: { id: params?.id },
     skip: !params?.id,
     ssr: false,
+  })
+
+  const { data: canEditBuyerOrgEdit } = useOrgPermission({
+    resourceCode: 'buyer_organization_edit',
   })
 
   /**
@@ -470,6 +475,7 @@ const OrganizationDetails: FunctionComponent = () => {
             variation="primary"
             isLoading={loadingState}
             onClick={() => handleUpdateOrganization()}
+            disabled={!canEditBuyerOrgEdit}
           >
             <FormattedMessage id="admin/b2b-organizations.organization-details.button.save" />
           </Button>