+++ react/components/CreateOrganizationButton/CreateOrganizationButton.tsx
@@ -32,6 +32,7 @@ import {
   B2B_BULK_IMPORT_TEMPLATE_LINK,
   B2B_BULK_IMPORT_BEST_PRACTICES_LINK,
 } from '../../utils/constants'
+import { useOrgPermission } from '../../hooks/useOrgPermission'
 
 const CreateOrganizationButton = () => {
   const { formatMessage, translate: t } = useTranslate()
@@ -42,6 +43,9 @@ const CreateOrganizationButton = () => {
   const { startBulkImport } = useStartBulkImport()
 
   const { data } = useBulkImportsQuery()
+  const { data: canEditBuyerOrgEdit } = useOrgPermission({
+    resourceCode: 'buyer_organization_edit',
+  })
 
   const handleUploadFinish = async (result: UploadFileData) => {
     if (result.status === 'success' && result.data?.fileData?.importId) {
@@ -56,6 +60,7 @@ const CreateOrganizationButton = () => {
         state={menuState}
         label={formatMessage(messages.new)}
         labelHidden={false}
+        disabled={!canEditBuyerOrgEdit}
         variant="primary"
       />
       <Menu state={menuState} aria-label="actions">